+++ bokeh/server/tests/usermodel_test.py
@@ -8,9 +8,9 @@
 from __future__ import absolute_import
 
 from . import test_utils
-from .. import models
 from ..app import bokeh_app
 from ..models import user
+from ..models.server_model import UnauthorizedException
 
 class TestUser(test_utils.BokehServerTestCase):
     def setUp(self):
@@ -19,15 +19,15 @@ class TestUser(test_utils.BokehServerTestCase):
 
     def test_cant_create_twice(self):
         user.new_user(self.client, 'test@test.com', 'mypassword', docs=[1, 2, 3])
-        self.assertRaises(models.UnauthorizedException, user.new_user,
+        self.assertRaises(UnauthorizedException, user.new_user,
                           self.client, 'test@test.com', 'mypassword')
 
     def test_auth_user(self):
-        self.assertRaises(models.UnauthorizedException,
+        self.assertRaises(UnauthorizedException,
                           user.auth_user,
                           self.client, 'test@test.com', 'mypassword')
         model = user.new_user(self.client, 'test@test.com', 'mypassword')
         assert model.username == 'test@test.com'
         model = user.auth_user(self.client, 'test@test.com', 'mypassword')
-        self.assertRaises(models.UnauthorizedException, user.auth_user,
+        self.assertRaises(UnauthorizedException, user.auth_user,
                           self.client, 'test@test.com', 'wrongpassword')
